requirements.txt
flask
flask-cors
numpy
gunicorn
tensorflow-cpu